+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Flag to enable debug mode and non-secure session cookie.
 - Show bitmap coverage information among overall results.
 
 ### Fixed